+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- fixed: Disabling pin-login while in duress mode disables pin-login for the main login.
+- fixed: `loginWithPassword` bug disabled duress mode when doing an online login.
+
 ## 2.27.4 (2025-05-13)
 
 - fixed: Disable duress pin mistakenly disables pin-login for entire account.